+++ orm/src/main/java/org/hibernate/tool/internal/export/lint/InstrumentationDetector.java
@@ -1,4 +1,4 @@
-package org.hibernate.tool.hbmlint.detector;
+package org.hibernate.tool.internal.export.lint;
 
 import org.hibernate.MappingException;
 import org.hibernate.boot.Metadata;
@@ -7,9 +7,6 @@ import org.hibernate.cfg.Environment;
 import org.hibernate.engine.spi.Managed;
 import org.hibernate.mapping.PersistentClass;
 import org.hibernate.mapping.Property;
-import org.hibernate.tool.internal.export.lint.EntityModelDetector;
-import org.hibernate.tool.internal.export.lint.Issue;
-import org.hibernate.tool.internal.export.lint.IssueCollector;
 
 public class InstrumentationDetector extends EntityModelDetector {
